# src/config/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "lib-quality",
    "description": "Collects data from GitHub, consolidate, and make them available through a REST api.",
    "contact": {
      "email": "juniorug@gmail.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:8080",
  "basePath": "/",
  "tags": [
    {
      "name": "projects",
      "description": "Everything about Projects in GitHub"
    }
  ],
  "schemes": ["http", "https"],
  "paths": {
    "/start": {
      "get": {
        "tags": ["projects"],
        "summary": "Pull issue data from github, consolidate it and make available for consultation",
        "produces": ["application/json"],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/projects": {
      "get": {
        "tags": ["projects"],
        "summary": "Get info for all registered projects from github",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/projects/{id}": {
      "get": {
        "tags": ["projects"],
        "summary": "Get info from a specific project by given id ",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "type": "number",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/projects/name/{name}": {
      "get": {
        "tags": ["projects"],
        "summary": "Get info from a specific project by given name from github",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "name",
            "type": "string",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  }
}